✨ feat: 카트 계산 로직을 커스텀 훅 스타일로 분리
- src/basic/hooks/useCart.js 파일 생성
- handleCalculateCartStuff 로직을 커스텀 훅으로 추출
- doUpdatePricesInCart 로직도 훅 내부로 이동
- main.basic.js에서 해당 훅 import하여 사용하도록 변경
